feat: add typescript, vite and refactoring of the code
