--- frontend/src/SearchPage/SearchPageStyles.js
@@ -1,16 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-const useStyles = makeStyles((theme) => ({
-  textField: {
-    "& .MuiOutlinedInput-root": {
-      "& fieldset": {
-        border: "none",
-      },
-    },
-  },
-  buttonStyle: {
-    borderColor: 'white',
-  },
-}));
-
-export default useStyles;
