+++ frontend/src/components/SignInButton/constants.ts
@@ -1,0 +1,1 @@
+export const SIGN_IN = "Sign in";
