+++ frontend/src/components/SignInButton/styles.css
@@ -1,0 +1,9 @@
+.text-button-style {
+  font-family: "Roboto";
+  font-size: 24px;
+  color: #A61700;
+}
+
+.button-style {
+  border-color: white;
+}
