+++ frontend/src/pages/SearchPage/SearchPage.css
@@ -1,26 +1,3 @@
-body {
-  margin: 0;
-  padding: 0;
-  background-color: #FFBA73;
-
-}
-
-html {
-  margin: 0;
-  padding: 0;
-}
-
-h1 {
-  display: flex;
-  font-family: 'Righteous';
-  font-style: normal;
-  font-weight: 400;
-  font-size: 85px;
-  line-height: 106px;
-  margin-block-start: 0;
-  margin-block-end: 0;
-}
-
 .title {
   display: flex;
   flex-direction: column;
@@ -45,7 +22,7 @@ h1 {
   height: 100vh;
   justify-content: space-around;
   flex-direction: column;
-  background-image: url(Bubble.svg);
+  background-image: url('/assets/images/bubble.svg');
   background-repeat: no-repeat;
 }
 
@@ -85,7 +62,7 @@ h1 {
   display: flex;
   width: 40%;
   height: 40vh;
-  background-image: url(Limon.svg);
+  background-image: url('/assets/images/limon.svg');
   background-repeat: no-repeat;
   background-size: contain;
 }
@@ -99,4 +76,16 @@ h1 {
 .space {
   display: flex;
   flex-basis: 40%;
+}
+
+.text-field {
+  & .MuiOutlinedInput-root {
+    & fieldset {
+      border: none;
+    }
+  }
+}
+
+.button-style {
+  border-color: white;
 }
