+++ frontend/src/pages/SearchPage/constants.ts
@@ -1,2 +1,1 @@
-export const SIGN_IN = "Sign in";
 export const SUPPORT = "Support";
